Create integration test to ensure rate limiter work
