+++ .env.example
@@ -1,5 +1,5 @@
 # APP
-BLINK_APP_DEFAULT_PREFETCH_COUNT=50
+BLINK_APP_DEFAULT_PREFETCH_COUNT=30
 
 # Disale New Relic locally
 NEW_RELIC_NO_CONFIG_FILE=true
